put linkTagsTo and unlinkActivitiesFrom in their own file
